Centralize modular scale logic
